Updated Tomcat & Java Cypher Server
Updated the cypher server implement multithreading in this version,
there's still a few bugs in the application although.

The tomcat client has also been updated, some adjustments to the
multithreading and the way messages and queues are handled.

Structure of the project changed again to one package only in the
project source files.